Created two Custom Hooks for fetching movies and add movies inside Local Storage in  App.js file
